src/requirements.txt
Flask
flask-mysqldb
flask-bcrypt
flask-login
Flask-SocketIO
mysql-connector-python
ultralytics
deep_sort_realtime
opencv-python
numpy
redis